remove unnecessary cast

+++ models/demos/deepseek_v3_d_p/tests/op_unit_tests/test_dispatch_combine_l1_small_semaphores.py
@@ -86,7 +86,7 @@ def run_dispatch_op(mesh_device, use_l1_small):
     )
     # #44928: dispatch consumes UINT16 indices directly from moe_grouped_topk.
     tt_indices = ttnn.from_torch(
-        indices.to(torch.int16),
+        indices,
         mesh_mapper=mesh_mapper,
         layout=ttnn.ROW_MAJOR_LAYOUT,
         device=mesh_device,
